add card functionality posts to server and fetches newest cards
